refactor : money값들 String에서 Long으로 변경(BIGINT)

+++ src/main/java/com/oopsw/clario/dto/dashboard/Top3CategoryDTO.java
@@ -10,5 +10,5 @@ import lombok.*;
 public class Top3CategoryDTO {
     private int memberId;
     private String categoryName;
-    private int categoryMoney;
+    private Long categoryMoney;
 }
